refactor(tests): extract style-contract scanner layer into style-contract-lib
Split the 1037-line renderer style contract test into suites-only
(462 lines) plus a reusable scanner/vocabulary module. Behavior
verified identical via a differential harness against HEAD; 25/25
contract tests pass.
